requirements.txt
aiogram>=3.0.0
apscheduler
sqlalchemy[aiosqlite]
cachetools
aiohttp
beautifulsoup4
lxml
pytz
pandas
google-auth
google-auth-oauthlib
python-dotenv
